Fix borders of _targetRegion not being cleared completely
Using cv::rectangle doesn't work because the line is _halfMatchSize
but is centered, not placed within, the Rect parameter. The
documentation is not clear on how it rounds everything, and on top
of that, I have seen the rectangles missing corner pixels in some
cases. Therefor this is the most most reliable solution, even if
it is a bit uglier.

Fixes #17